logger/requirements.txt
pywin32
winregistry
watchdog
pynput